Get further in pthread startup.
Most of the fixes here are 32/64 issues. In any event, we get past
previous mmap issues and fail at the clone which we don't support yet.

Reviewed-on: http://nxm.coreboot.org/review/105
Tested-by: Jenkins QA

+++ sys/src/nxm/port/sysseg.c
@@ -64,7 +64,7 @@ ibrk(uintptr addr, int seg)
 {
 	Segment *s, *ns;
 	uintptr newtop, rtop;
-	long newsize;
+	uintptr newsize;
 	int i, mapsize;
 	Pte **map;
 	uintmem pgsz;
